fix login and review
