+++ src/Snap.Hutao/Snap.Hutao/UI/Xaml/View/Dialog/UserAccountPasswordDialog.xaml
@@ -5,7 +5,7 @@
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:shuxm="using:Snap.Hutao.UI.Xaml.Markup"
-    Title="账号密码登录"
+    Title="{shuxm:ResourceString Name=ViewDialogUserAccountPasswordTitle}"
     CloseButtonText="{shuxm:ResourceString Name=ContentDialogCancelCloseButtonText}"
     DefaultButton="Close"
     IsPrimaryButtonEnabled="{x:Bind IsLoginEnabled, Mode=OneWay}"
@@ -17,13 +17,13 @@
         <TextBox
             KeyDown="OnTextKeyDown"
             KeyUp="OnTextKeyUp"
-            PlaceholderText="请输入账号"
+            PlaceholderText="{shuxm:ResourceString Name=ViewDialogUserAccountPasswordAccountHint}"
             Text="{x:Bind Account, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}"/>
         <PasswordBox
             KeyDown="OnTextKeyDown"
             KeyUp="OnTextKeyUp"
             Password="{x:Bind Password, Mode=TwoWay}"
             PasswordChar="*"
-            PlaceholderText="请输入密码"/>
+            PlaceholderText="{shuxm:ResourceString Name=ViewDialogUserAccountPasswordPasswordHint}"/>
     </StackPanel>
 </ContentDialog>
